// projects/forms/src/lib/model/input-type.ts
export type InputType =
  | 'text'
  | 'tel'
  | 'url'
  | 'color'
  | 'date'
  | 'password'
  | 'number'
  | 'checkbox'
  | 'email'
  | 'button'
  | 'hidden'
  | 'reset'
  | 'month'
  | 'week'
  | 'range'
  | 'image'
  | 'search'
  | 'time'
  | 'radio'
  | 'file'
  | 'datetime-local'
  | 'submit';